changed to dialogFragment
