+++ code/app/src/main/java/com/example/projecteventlotteryapp/CreateEventDialogFragment.java
@@ -1,0 +1,54 @@
+package com.example.projecteventlotteryapp;
+
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * A simple {@link DialogFragment} subclass.
+ * Use the {@link CreateEventDialogFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class CreateEventDialogFragment extends DialogFragment {
+    private static final String ARG_PARAM1 = "param1";
+    private static final String ARG_PARAM2 = "param2";
+
+    private String mParam1;
+    private String mParam2;
+
+    public CreateEventDialogFragment() {
+        // Required empty public constructor
+    }
+
+    public static CreateEventDialogFragment newInstance(String param1, String param2) {
+        CreateEventDialogFragment fragment = new CreateEventDialogFragment();
+        Bundle args = new Bundle();
+        args.putString(ARG_PARAM1, param1);
+        args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            mParam1 = getArguments().getString(ARG_PARAM1);
+            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        // Inflate the same layout as before
+        return inflater.inflate(R.layout.fragment_create_event, container, false);
+    }
+}
